ai_services/stroke_image/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
numpy==1.24.3
pillow==10.0.0
tensorflow==2.13.0
